# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/reviews/{id}": {
            "put": {
                "description": "Updates an existing review in the database.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reviews"
                ],
                "summary": "Update Review",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Review ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated review data",
                        "name": "review",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Review"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Review"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.JSONResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.JSONResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.NullableFloat64": {
            "type": "object",
            "properties": {
                "valid": {
                    "type": "boolean"
                },
                "value": {
                    "type": "number"
                }
            }
        },
        "models.Review": {
            "type": "object",
            "required": [
                "app",
                "sentiment",
                "translated_review"
            ],
            "properties": {
                "app": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "sentiment": {
                    "type": "string"
                },
                "sentiment_polarity": {
                    "$ref": "#/definitions/models.NullableFloat64"
                },
                "sentiment_subjectivity": {
                    "$ref": "#/definitions/models.NullableFloat64"
                },
                "translated_review": {
                    "type": "string"
                }
            }
        },
        "utils.JSONResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "status": {
                    "type": "string"
                }
            }
        }
    }
}